Add local procedure calls.
Each top-level procedure and any procedures local to it get registers allocated as a single block of code.  This avoids the need for a commmon call protocol with callee- or caller-saved registers.
This does not apply to any internal recursive procedures, which will need to get registers allocated separately.

Closed-over variables are not yet handled.